requirements.txt
langchain
langchain-google-genai
langchain-community
faiss-cpu
python-dotenv
google-generativeai